--- InMemoryImpOutputDestination.java
@@ -1,10 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class InMemoryImpOutputDestination implements OutputDestination{
-    private List <String> data = new ArrayList<>();
-    public List<String> writeData() {
-        return data;
-    }
-}
-
